backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "task-tracker"
version = "0.1.0"
description = "Task Tracker - Production-ready task management application with FastAPI"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    # Core framework
    "fastapi>=0.121.2",
    "uvicorn[standard]>=0.38.0",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.12.0",
    "pydantic[email]>=2.12.4",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "aiosqlite>=0.20.0",
    # Database
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "psycopg2-binary>=2.9.0",
    "asyncpg>=0.29.0",
    # Authentication
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.0",
    # Caching and Queue
    "redis>=5.0.0",
    "celery>=5.3.0",
    # Monitoring and Observability
    "prometheus-client>=0.19.0",
    "prometheus-fastapi-instrumentator>=7.0.0",
    # Utilities
    "requests>=2.32.5",
    "greenlet>=3.2.4",
]

[dependency-groups]
dev = [
    "black>=25.11.0",
    "pytest>=9.0.1",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.14.5",
    "httpx>=0.27.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
asyncio_mode = "auto"

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.ruff]
line-length = 100
target-version = "py312"
select = ["E", "F", "W"]
ignore = ["E501"]  # Line too long (handled by black)

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]  # Imported but unused
